🐛 fix: fix data type amount and balance done

+++ controllers/transactions.controller.go
@@ -14,7 +14,7 @@ import (
 )
 
 type inputTransaction struct {
-	Amount      uint      `json:"amount" form:"amount"`
+	Amount      uint64    `json:"amount" form:"amount"`
 	CategoryID  uint      `json:"category_id" form:"category_id"`
 	WalletID    uint      `json:"wallet_id" form:"wallet_id"`
 	Type        string    `json:"type" form:"type"`
